+++ back/boxtribute_server/models/definitions/classic_location.py
