+++ back/test/endpoint_tests/test_classic_location.py
